Fix packet-lapd.c:186: warning: comparison between signed and unsigned.
svn path=/trunk/; revision=54014

+++ epan/dissectors/packet-lapd.c
@@ -183,7 +183,7 @@ fill_lapd_byte_state(lapd_byte_state_t *ptr, enum lapd_bitstream_states state, c
 	ptr->bit_offset = bit_offset;
 	ptr->ones = ones;
 
-	ptr->data_len = MIN(sizeof(ptr->data), data_len);
+	ptr->data_len = MIN((int)sizeof(ptr->data), data_len);
 	memcpy(ptr->data, data, ptr->data_len);
 }
